requirements.txt
astroid==2.15.0
blingfire==0.1.8
CacheControl==0.12.11
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==39.0.2
dill==0.3.6
filelock==3.10.0
firebase-admin==6.1.0
Flask==2.2.3
google-api-core==2.11.0
google-api-python-client==2.82.0
google-auth==2.16.2
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.10.0
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.59.0
grpcio==1.51.3
grpcio-status==1.51.3
gunicorn==20.1.0
httplib2==0.22.0
huggingface-hub==0.13.2
idna==3.4
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
lazy-object-proxy==1.9.0
MarkupSafe==2.1.2
mccabe==0.7.0
mpmath==1.3.0
msgpack==1.0.5
networkx==3.0
numpy==1.24.2
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
platformdirs==3.1.1
proto-plus==1.22.2
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pylint==2.17.0
pyparsing==3.0.9
python-crfsuite==0.9.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
pyvi==0.1.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
sentencepiece==0.1.97
six==1.16.0
sklearn-crfsuite==0.3.6
sympy==1.11.1
tabulate==0.9.0
threadpoolctl==3.1.0
tokenizers==0.13.2
tomlkit==0.11.6
tqdm==4.65.0
transformers==4.27.1
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
Werkzeug==2.2.3
wrapt==1.15.0
